requirements.txt
streamlit>=1.24.0
streamlit-nightly
streamlit-audiorec
openai
requests
assemblyai
PyPDF2
pdfminer.six
langchain 
faiss-cpu
python-docx
pydantic




